Enable antialiasing by default on QML scenes

+++ src/core/qgssggeometry.h
@@ -24,12 +24,12 @@ class QgsSGGeometry : public QSGNode
 {
   public:
     QgsSGGeometry();
-    QgsSGGeometry( const QgsGeometry &geom, const QColor &color, int width, const QgsRectangle visibleExtent, double scaleFactor );
+    QgsSGGeometry( const QgsGeometry &geom, const QColor &color, float width, const QgsRectangle visibleExtent, double scaleFactor );
 
   private:
     void applyStyle( QSGGeometryNode *geomNode );
 
-    static QSGGeometry *qgsPolylineToQSGGeometry( const QgsPolylineXY &line, int width, const QgsRectangle visibleExtent, double scaleFactor );
+    static QSGGeometry *qgsPolylineToQSGGeometry( const QgsPolylineXY &line, float width, const QgsRectangle visibleExtent, double scaleFactor );
     static QSGGeometry *qgsPolygonToQSGGeometry( const QgsPolygon *polygon, const QgsRectangle visibleExtent, double scaleFactor );
 
     QSGFlatColorMaterial mMaterial;
